worker/pyproject.toml
[tool.poetry]
name = "worker"
version = "0.1.0"
description = "Speech-to-code worker."
authors = ["Niels Horn <niels@horn.ninja>"]
license = "MIT license"
readme = "README.md"

[tool.poetry.scripts]
start = "worker.main:start"

[tool.poetry.dependencies]
python = "~3.11"
speech-to-code = {path = ".."}
azure-storage-queue = "^12.9.0"
azure-storage-blob = "^12.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
